Get a lot of IRC details out of places.
Events no longer act as parsers for raw messages; the raw message isn't
even required.  Instead, the particular message handlers in the client
are responsible for building events using their various APIs, as it
should've been in the first place goddamn.

This got rid of a lot of wacky cruft already.